+++ src/AnimatedTabBar.tsx
@@ -2,8 +2,8 @@ import React, { useMemo, useCallback, useEffect, useRef } from 'react';
 import type { LayoutChangeEvent } from 'react-native';
 import { useSafeArea } from 'react-native-safe-area-context';
 import { AnimatedTabBarView } from './AnimatedTabBarView';
-import { useStableCallback } from './utilities';
-import { useTabBarVisibility } from './hooks';
+import { interpolate } from './utilities';
+import { useTabBarVisibility, useStableCallback } from './hooks';
 import type { PresetEnum } from './presets';
 import type { AnimatedTabBarProps } from './types';
 import Animated, {
@@ -15,9 +15,6 @@ import Animated, {
   onChange,
 } from 'react-native-reanimated';
 import { useValue } from 'react-native-redash/lib/module/v1';
-
-// @ts-ignore fallback for react-native-reanimated v2
-const interpolate = Animated.interpolateNode ?? Animated.interpolate;
 
 interface Route {
   name: string;
