+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export { useTabBarVisibility } from './useTabBarVisibility';
 export { useTabBarItemFocusTransition } from './useTabBarItemFocusTransition';
 export { useTabBarItemSpacing } from './useTabBarItemSpacing';
+export { useStableCallback } from './useStableCallback';
