+++ src/presets/bubble/item/BubbleTabBarItem.tsx
@@ -4,12 +4,11 @@ import Animated from 'react-native-reanimated';
 import { interpolateColor, useValue } from 'react-native-redash/lib/module/v1';
 // @ts-ignore 😞
 import isEqual from 'lodash.isequal';
+import { interpolate } from '../../../utilities';
 import type { BubbleTabBarItemProps } from '../types';
 import { styles } from './styles';
 
 const { add } = Animated;
-// @ts-ignore fallback for react-native-reanimated v2
-const interpolate = Animated.interpolateNode ?? Animated.interpolate;
 
 const BubbleTabBarItemComponent = ({
   animatedFocus,
