+++ src/utilities/noop.ts
@@ -1,0 +1,1 @@
+export const noop = () => {};
